REFACTOR: Consolidate TS code for communication with backend.
