+++ packages/frontend/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./automerge";
+export * from "./rpc";
