+++ packages/frontend/src/components/rich_text_editor.tsx
@@ -9,7 +9,7 @@ import { type Command, EditorState, Plugin, type Transaction } from "prosemirror
 import { EditorView } from "prosemirror-view";
 
 import { createEffect, onCleanup } from "solid-js";
-import { useDocHandleReady } from "../util/automerge_solid";
+import { useDocHandleReady } from "../api/automerge";
 
 import "prosemirror-view/style/prosemirror.css";
 import "./rich_text_editor.css";
